packages/rag/pyproject.toml
[project]
name = "rag"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["pandas<3", "lancedb==0.17", "python-dotenv"]
